unset var fix

+++ ci/basic_nightly_test.sh
@@ -6,15 +6,19 @@ set -euxo pipefail
 # temp test
 zig version
 
-# Remove everything in this dir except the tar and ci folder.
-# We want to test like a user who would have downloaded the release, so we clean up all files from the repo checkout.
-to_delete=$(ls | grep -v "roc_nightly.*tar\.gz"  | grep -v "^ci$")
+# if to prevent unset vars errror
+if [ -n "$(ls | grep -v "roc_nightly.*tar\.gz"  | grep -v "^ci$")" ]; then
+  
+  # Remove everything in this dir except the tar and ci folder.
+  # We want to test like a user who would have downloaded the release, so we clean up all files from the repo checkout.
+  to_delete=$(ls | grep -v "roc_nightly.*tar\.gz"  | grep -v "^ci$")
 
-for file_or_dir in $to_delete
-do
+  for file_or_dir in $to_delete
+  do
     echo "Removing: $file_or_dir" 
     rm -rf "$file_or_dir"
-done
+  done
+fi
 
 # decompress the tar
 ls | grep "roc_nightly.*tar\.gz" | xargs tar -xzvf
